==> notebooks/1.0-extract-metadata.ipynb <==
!pip3 install pillow

from PIL import Image
from PIL.ExifTags import TAGS
import os
import pandas as pd

# extract other basic metadata
info_dict = {
    "Filename": image.filename,
    "Image Size": image.size,
    "Image Height": image.height,
    "Image Width": image.width,
    "Image Format": image.format,
    "Image Mode": image.mode,
    "Image is Animated": getattr(image, "is_animated", False),
    "Frames in Image": getattr(image, "n_frames", 1)
}

for label,value in info_dict.items():
    print(f"{label:25}: {value}")

# extract EXIF data
# exifdata = image.getexif()
# print(exifdata)

def create_metadata(image_directory,extra_file):
    '''
    This Function downloads the csv file with metadata for image folder provided
    Parameters:
    ---------------------
    image_directory: string
        Directory of Images where they are all stored.
    extra_file: string 
        File Path for the file which you want to merge with the Dataset for this example it is annotations file
        
    '''
    # assign directory
    annotations = pd.read_csv(extra_file)
    # iterate over files in
    # that directory
    info_dict = {}
    for filename in os.listdir(image_directory):
        print("Reading file:",filename)
        f = os.path.join(image_directory, filename)
        imagename = f
        info_dict[filename] = {}
        # # read the image data using PIL
        image = Image.open(imagename)
        info_dict[filename]['image_size']  = image.size
        info_dict[filename]['image_Height']  = image.height
        info_dict[filename]['image_Width']  = image.width
        info_dict[filename]['image_Format']  = image.format
        info_dict[filename]['image_mode'] = image.mode
        info_dict[filename]['image_Animated']  = getattr(image, "is_animated", False)
        info_dict[filename]['image_Frames']  = getattr(image, "n_frames", 1)
        print("Inserted in DataFrame")

    df_final = pd.DataFrame.from_dict(info_dict)

    df = df_final.T
    df.reset_index(inplace= True)
    print("Merging Extra Files")
    df = pd.merge(df,annotations,left_on = 'index', right_on = 'image_id',how = 'left')
    df.drop(columns=["id", "index"], inplace=True)
    df.to_csv('../data/raw/MetaData.csv')
    print("***************************")
    print(df.head())

#run when file is directly executed
if __name__ == '__main__':
    create_metadata('/Users/lizhijie_1/Downloads/Aircraft-Detection/images/','/Users/lizhijie_1/Downloads/Aircraft-Detection/annotations.csv')
    print("done")

os.getcwd()
