==> notebooks/2.0-initial-pandas-profilling.ipynb <==
import numpy as np
import pandas as pd
from pandas_profiling import ProfileReport
import ast

df = pd.read_csv('../data/raw/MetaData.csv');

geo_column = df.loc[:,"geometry"].apply(lambda col: ast.literal_eval(col))
geo_column


xmin = geo_column.apply(lambda col: col[0][0])
xmax = geo_column.apply(lambda col: col[2][0])
ymin = geo_column.apply(lambda col: col[0][1])
ymax = geo_column.apply(lambda col: col[2][1])

width = geo_column.apply(lambda col: col[2][0]-col[0][0])
width

height = geo_column.apply(lambda col: col[2][1]-col[0][1])
height

df_plane = df.assign(Xmin=xmin, Xmax=xmax, Ymin=ymin, Ymax=ymax, width=width, height=height)
df_plane.drop(columns=['image_size', 'geometry'], inplace=True)
df_plane.drop(df.columns[df.columns.str.contains('unnamed',case = False)],axis = 1, inplace = True)
df_plane

df_plane.to_csv("../data/processed/airplane.csv")

profile = ProfileReport(df_plane, title="Pandas Profiling Report")
profile.to_file("../reports/airplane_data_profiling.html")
