==> notebooks/4.0-function.ipynb <==
import pandas as pd
from PIL import Image, ImageDraw
import os
from IPython.display import display, Image


df = pd.read_csv("../data/processed/image_planes_num.csv")
plane = pd.read_csv("../data/processed/airplane.csv")

df

def count_airplanes_in_given_image(image_id) -> int:
   # use 014de911-7810-4f7d-8967-3e5402209f4a.jpg as example
   ## Input image_id and compute how many aircraft contains in the given image
   ## return integer
   ## if image_id has no found responding image, throw error: no such image
   df1 = df[df['index']==True]
   return df1.loc[: ,"count"].sum()

count_airplanes_in_given_image("3")

count_airplanes_in_given_image("014de911-7810-4f7d-8967-3e5402209f4a.jpg")

def return_images_with_maximum_airplanes(number_of_image):
    withSum = df.groupby(["index"])["count"].sum().reset_index(name="sum")
    df2 = withSum.sort_values(by="sum", ascending=False)
    return df2.head(number_of_image);

return_images_with_maximum_airplanes(10)

def return_images_with_given_number_of_aircraft(contain_aircraft_number, limit_of_image) -> str:
    ## Input integer of how many aircraft contains 
    # Iterate through all images, find images that contains given number of aircraft
    # display up to 5 of them 
    # return number of these image

    # if no image contain such number of aircraft, return none
    # if input < 0 return error

    """
    a schema based on each image_id
    Image_id  aircraft_number  has_truncated  
                0 or [number]   0 or [truncated]
    """
    withSum = df.groupby(["index"])["count"].sum().reset_index(name="sum")
    df3 = withSum[withSum['sum'] == contain_aircraft_number]
    return df3.head(limit_of_image)

return_images_with_given_number_of_aircraft(20, 5)

def return_all_images_with_truncated_aircraft(limit_of_images):
    """ Input 5 would reutrn at most 5 image_id of truncated images
    """
    hasTruncated = df[df['class'] == "Truncated_airplane"]
    df4 = hasTruncated.sort_values(by="count", ascending=False)
    return df4.head(limit_of_images)

return_all_images_with_truncated_aircraft(10)

def return_truncated_airplane_info(image_id):
    """if no truncated, throw errors"""

    

def has_aircraft_in_given_x_y_coordinate(image_id, x_loc, y_loc):
    ## Input image_id and x, y coordinate  x (0, 2560)
    #  if the location of the image has a aircraft, return true, else return false

    """if false, return [] all other geometry contains airplane
    
    if true return image which has bounds box on entered location"""
    img = Image.open(r"/Users/lizhijie_1/Downloads/Aircraft-Detection/images/"+str(image_id))
    image = ImageDraw.Draw(img)
    
    test = plane[plane["image_id"]==image_id]
    test.reset_index(drop=True, inplace=True)

    coordinate = []
    for i in range(len(test.index)):
        tuple = (test.at[i, "Xmin"], test.at[i,"Ymin"], test.at[i, "Xmax"], test.at[i, "Ymax"])
        coordinate.append(tuple)
        if(x_loc>=tuple[0] and x_loc <= tuple[2] and 
            y_loc >= tuple[1] and y_loc <= tuple[3]):
            rect = image.rectangle(tuple, outline="red", width=5)
            img.show()
            return "Found"
        i+=1

    return "No airplane in given coordinate, here is all location that has airplanes", coordinate;


res = has_aircraft_in_given_x_y_coordinate("5c9e817a-dc4b-42ab-952c-3128e2de12e8.jpg",1600,500)
res

res1 = has_aircraft_in_given_x_y_coordinate("5c9e817a-dc4b-42ab-952c-3128e2de12e8.jpg", 2000, 1600)
res1

test1 = test.sort_values("plane_size", ascending=True)
test1.reset_index(drop=True, inplace=True)
test1.head(5)
# len(test.index)

im = Image.open(r"/Users/lizhijie_1/Downloads/Aircraft-Detection/images/5c9e817a-dc4b-42ab-952c-3128e2de12e8.jpg")
im.show()

def display_biggest_aircraft(image_id, limit_of_number):
    """if true, return list of (xmin, ymin, xmax, ymax) of all big airplane"""


    test = plane[plane["image_id"]==image_id]
    test.insert(0, "plane_size", test["width"]+test["height"])
    test1 = test.sort_values("plane_size", ascending=False)
    test1.reset_index(drop=True, inplace=True)

    img = Image.open(r"/Users/lizhijie_1/Downloads/Aircraft-Detection/images/"+str(image_id))
    image = ImageDraw.Draw(img)

    coordinate = []
    if(len(test1.index) >= limit_of_number):
        maximum = limit_of_number
    else: 
        maximum = len(test1.index)
    for i in range(maximum):
        tuple = (test1.at[i, "Xmin"], test1.at[i,"Ymin"], test1.at[i, "Xmax"], test1.at[i, "Ymax"])
        coordinate.append(tuple)
        rect = image.rectangle(tuple, outline="red", width=5)
        # print("current i = ",i)
        i+=1

    img.show()

    return coordinate; 

display_biggest_aircraft("5c9e817a-dc4b-42ab-952c-3128e2de12e8.jpg",30)

def display_smallest_aircraft(image_id, limit_of_number):

    test = plane[plane["image_id"]==image_id]
    test.insert(0, "plane_size", test["width"]+test["height"])
    test1 = test.sort_values("plane_size", ascending=True)
    test1.reset_index(drop=True, inplace=True)

    img = Image.open(r"/Users/lizhijie_1/Downloads/Aircraft-Detection/images/"+str(image_id))
    image = ImageDraw.Draw(img)

    coordinate = []
    if(len(test1.index) >= limit_of_number):
        maximum = limit_of_number
    else: 
        maximum = len(test1.index)
    for i in range(maximum):
        tuple = (test1.at[i, "Xmin"], test1.at[i,"Ymin"], test1.at[i, "Xmax"], test1.at[i, "Ymax"])
        coordinate.append(tuple)
        rect = image.rectangle(tuple, outline="red", width=5)
        print("current i = ",i)
        i+=1

    img.show()

    return coordinate; 

image1 = "5c9e817a-dc4b-42ab-952c-3128e2de12e8.jpg"
display_smallest_aircraft(image1, 5)
