==> notebooks/5.0-model-card-toolkit.ipynb <==
import pandas as pd
import seaborn as sns
import uuid
import verifyml.model_tests.utils as utils

from datetime import date
from sklearn.datasets import load_breast_cancer
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.metrics import plot_roc_curve
from sklearn.model_selection import train_test_split
